+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # CHANGELOG
 
-## 0.0.1
+## 0.0.3
 
-- changed stopPractice shortcut to "cmd+shift+q"
+- Added a "Configure Keybinds for Retype" command as a Ctrl+Shift+P menu option
+- Added "Reset Current Session" functionality
+- Fixed a bug where ` and ' were recognised as the same character
 
 ## 0.0.2
 
 - Changed the start typing test to use highlight instead of cursor
 - Fixed a bug where newlines could be typed twice
 - Changed startPractice and stopPractice to a togglePractice command, now using "Ctrl+`"
+
+## 0.0.1
+
+- Changed stopPractice shortcut to "cmd+shift+q"